+++ ConsoleApplication1/ConsoleApplication1/ConsoleApplication1.vcxproj
@@ -128,6 +128,10 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="ConsoleApplication1.cpp" />
+    <ClCompile Include="MokiniuProcessing.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="MokiniuProcessing.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
