+++ ConsoleApplication1/ConsoleApplication1/x64/Release/ConsoleA.f92d470b.tlog/Cl.items.tlog
@@ -1,1 +1,2 @@
 C:\Users\skirm\Documents\GitHub\objektinis1\ConsoleApplication1\ConsoleApplication1\ConsoleApplication1.cpp;C:\Users\skirm\Documents\GitHub\objektinis1\ConsoleApplication1\ConsoleApplication1\x64\Release\ConsoleApplication1.obj
+C:\Users\skirm\Documents\GitHub\objektinis1\ConsoleApplication1\ConsoleApplication1\MokiniuProcessing.cpp;C:\Users\skirm\Documents\GitHub\objektinis1\ConsoleApplication1\ConsoleApplication1\x64\Release\MokiniuProcessing.obj
