+++ ConsoleApplication1/ConsoleApplication1/x64/Release/ConsoleApplication1.log
@@ -1,6 +1,8 @@
-﻿  Generating code
-  0 of 459 functions ( 0.0%) were compiled, the rest were copied from previous compilation.
-    0 functions were new in current compilation
-    0 functions had inline decision re-evaluated but remain unchanged
+﻿  ConsoleApplication1.cpp
+C:\Users\skirm\Documents\GitHub\objektinis1\ConsoleApplication1\ConsoleApplication1\ConsoleApplication1.cpp(34,20): warning C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data
+C:\Users\skirm\Documents\GitHub\objektinis1\ConsoleApplication1\ConsoleApplication1\ConsoleApplication1.cpp(68,15): warning C4244: 'argument': conversion from 'time_t' to 'unsigned int', possible loss of data
+  Generating code
+  Previous IPDB and IOBJ mismatch, fall back to full compilation.
+  All 459 functions were compiled because no usable IPDB/IOBJ from previous compilation was found.
   Finished generating code
   ConsoleApplication1.vcxproj -> C:\Users\skirm\Documents\GitHub\objektinis1\ConsoleApplication1\x64\Release\ConsoleApplication1.exe
